mds/snap: Fix snapshot read
The snapshot lookup was returning the head inode data
instead of the snapshotted inode data. This was because
the 'has_snapshots_in_range' was buggy and hence journal_cow_dentry
wasn't doing cow_old_inodes.

TODO: Need to audit all the apis in the SnapRealm and see if
it's considering the related_realms.
Since there is not dedicated realm for hardlinked inodes to converge
snapids and store. For now, building it on the fly when required and
not altering cached items is working fine. Need to see how to
improve this.

Signed-off-by: Kotresh HR <khiremat@redhat.com>