mds/kclient: Fix snaprealm split of inodes which had caps
If the inodes had caps exclusive caps issued, those needs
to considered while taking snapshot during snaprealm split.

This would be recorded in the list - inodes_with_caps of parent snaprealm.
When the snapshot is created at a directory 'x', new snaprealm is
created for x, during which the 'inodes_with_caps' list of parent
snaprealm is traversed and those inodes are moved to the new snaprealm
of 'x' if the directory 'x' is the ancestor.

For the hardlink case, until now global snaprealm was taking care of
this. Now finding the ancestor is using the referent inodes to find
if it's the ancestor.

This fixes the snapshot creation part at the backend

Signed-off-by: Kotresh HR <khiremat@redhat.com>