mds/Locker/kclient: Fix snapshot read - flush snap
The flush snap message from client (part of client_caps)
was failing on a hardlinked file as it couldn't find it's
corresponding snapshot id.

Consider the below directory tree.
    /
    - dir1
       - file1
       - file2
    - dir2
       - hl_file1
       - hl_file2

In the above case, the snapshot is taken on 'dir2', the real inode
is present under dir1. On flush snap msg, the real inode's snaprealm
would be / (root snaprealm) and it has no snaps. Hence the failure
to flush the snap. The fix is to aggregate all the related snaps
in this case.

This would fix the case of reading a file from the snapshot.

Signed-off-by: Kotresh HR <khiremat@redhat.com>