common/pick_address: support ipv6 in subnet check
Extend the is_addr_in_subnet function with support for ipv6 addresses.
To do this try to parse addresses as ipv6 after trying ipv4 and choose
the correct CEPH_PICK_ADDRESS family thereafter. Also extend the public
network iterator with support for ipv6 as well by using the correct
struct sockaddr6_in.

This fixes a reported HEALTH_ERR when the cluster uses only ipv6 with a
public and a cluster network. This was introduced in commit
1332cc428b3f ("osd: Report health error if OSD public address is not
within subnet") which never took an ipv6 only cluster into account and
relied on ipv4 only.

Fixes: https://tracker.ceph.com/issues/67517
Signed-off-by: Rouven Czerwinski <rouven@czerwinskis.de>