mgr/cephadm: Restore client file(s) default SELinux context type
When cephadm writes client files to hosts, the new files are created in
/tmp/cephadm-<fsid> and then moved to their target directory. This makes
the client files receive the usr_tmp_t context, which is different then
their original context.
Restorecon will restore the type portion of the security context to be
as defined in the SELinux policy for that file path.

Fixes: https://tracker.ceph.com/issues/69395
Signed-off-by: zaken <yonatan.zaken.ext@nokia.com>