doc: README.md - format "Troubleshooting"
Format "Troubleshooting" into its own section so that it doesn't confuse
readers of the vstart installation procedure.

This PR is part of a series of PRs meant to preserve the integrity of
the README.md file after some recent additions that break the flow of
the document.

This PR follows https://github.com/ceph/ceph/pull/61221.

Signed-off-by: Zac Dover <zac.dover@proton.me>