multi-mds/unlink: Fix referent_inode list
The referent inode which is removed as part of unlink was
getting added back at the commit phase unconditionally.
The referent inode should be added at the commit phase
for the link creation and not for unlink. The adding of
referent inode was moved to the commit phase to make the
rollback easier.

Signed-off-by: Kotresh HR <khiremat@redhat.com>