rados/monthrash: add PG_DEGRADED to ignorelist
Issue: mon/test_mon_osdmap_prune.sh is failing due to
PG_DEGRADED state.

Cause: We are writing an objects to a newly created PG before
it has enough OSDs in acting set to create enough replicas.
This is expected behaviour, thus the solution is to add the
warning to ignorelist.

Fixes: https://tracker.ceph.com/issues/69288

Signed-off-by: Shraddha Agrawal <shraddha.agrawal000@gmail.com>