cephadm: fix handling of ceph-exporter sock-dir
Fixes: https://tracker.ceph.com/issues/69475

It turns out the sock-dir for ceph-exporter only needs to
exist within the container, not on the host. Previous code,
including the validation function this commit removes
and previous patches trying to fix the ceph-exporter asok
file not appearing on the host, were all done assuming
it mattered what was on the host. This patch changes things
so all we do with the sock dir is mount it to /var/run/ceph/<fsid>
and don't worry about whether that dir exists on the host.
Additionally, the patch makes it so /var/run/ceph/<fsid> is
created during ceph-exporter deployment.

Signed-off-by: Adam King <adking@redhat.com>