qa: Moved setting up stretch mode from bash to python
Problem:
We enter stretch mode through a bash script
which is not homogenous with how we are performing
standalone tests. Moving it into Python allows for
code resusability, therefore, more neat and efficient.

Solution:
Rewrote the code in python, creating a separate helper file
such that the code can be reused.

Fixes: https://tracker.ceph.com/issues/68058

Signed-off-by: Kamoltat Sirivadhna <ksirivad@redhat.com>