mgr/cephadm: re-factoring the dependencies calculation code
currently, the dependency logic is duplicated between the different Service
classes and the module.py::_calc_daemon_deps function, which can lead
to issues such as BUGSs, difficulty in maintenance, and other problems
associated with duplicated code. In this change, we are consolidating
all the dependency logic into the Service subclasses to eliminate this
duplication. This way, we also force anybody creating a new Service to
think about its potential dependencies.

Fixes: https://tracker.ceph.com/issues/69021
Signed-off-by: Redouane Kachach <rkachach@ibm.com>