lmgr/cephadm: validate no duplicate groups and group vs. service id for nvmeof
This came as a recommendation from the nvmeof team because

1) users shouldn't have multiple nvmeof services with the same group

and

2) some users may take an existing nvmeof service spec and modify
the service id to point to another group but forget to update
the actual group parameter. This makes it look at a glance like
the two services are using different groups when they are in fact not.

Usually I'd write a spec migration for this, but if we were to find
users had two specs with duplicate groups or specs where the service
id doesn't match the group name, I don't think there's a clear path
of action to take to remedy the situation. Even just forcing the
service id of an nvmeof service to be pool_name.group_name I think
could cause a lot of confusion during an upgrade so I'm tentative
to do it.

Signed-off-by: Adam King <adking@redhat.com>