qa/suites/upgrade/telemetry-upgrade: ignore TELEMETRY_CHANGED warning
Even though we opt in right away after upgrading, which minimizes the chance
of this warning coming up, the cluster log still detects it sometimes and
fails the test. So, we can ignore it since it's expected for the warning
to come up for a few seconds in some cases.

Signed-off-by: Laura Flores <lflores@ibm.com>