crimson/net/io_handler: drop io_state_t::none
io_state_t::none was introduced to resemble async msgr and used by the
single-shard crimson msgr to mark that the connection is invisible by
any user during accepting. But it isn't simple in multi-shard crimson
msgr because the ordering is impossible to predict among the connection
user, connection and the msgr shards if they are in 3 different shards.

So drop io_state_t::none.

Fixes: https://tracker.ceph.com/issues/66606
Signed-off-by: Yingxin Cheng <yingxin.cheng@intel.com>