build: Link TBB into io_exerciser if present
libstdc++ uses TBB to implement the execution library if it is
available. If it's not present, we get a serial backend.

Currently, we aren't getting link errors in most cases because the
reference is optimized out, but when compiling with `-O0`, we hit a
missing symbol.

If we use more of the execution library, we'll reference TBB in ways
that don't optimize out.

As such, test if TBB is available. If so, link against it.

See https://gcc.gnu.org/onlinedocs/libstdc++/manual/status.html#status.iso.2017
for more information.

Signed-off-by: Adam Emerson <aemerson@redhat.com>