osd/recovery_types: Introduce is_existing_begin
BackfillInterval contains all objects in the range of [begin, end).
If an object was deleted in this range, our BackfillInterval::begin/end
markers would remain, but the objects would not be contained in
BackfillInterval::objects map.

Before adding begin to the stats, which would require obtaining its
obc later on. Make sure that it actaully exists.

Possible RCA for: https://tracker.ceph.com/issues/56764

Signed-off-by: Matan Breizman <mbreizma@redhat.com>