pybind/mgr/cephadm: fix issue with multiple nfs clusters on the same port
Currently even if ingress and virtual_ip are used, the port_in_use
check in cephadm fails when the same port is used. This PR fixes
the issue by ensuring that the daemon spec includes the virtual
ip address for the endpoint to be checked is included so it
doesn't default to checking `0.0.0.0:port` only.

Fixes: https://tracker.ceph.com/issues/69744
Signed-off-by: Omid Yoosefi <omidyoosefi@ibm.com>