cephadm: use parsed_container_mem_usage in cephadm.py
Replace the use of _parse_mem_usage and related command calls with
parsed_container_mem_usage.

This needs a lot of tests to be updated so that the call functions that
now originate in `container_engines.py` get mocked. funkypatch handles
most of the extra work for us - but it needs to be added to many tests.

Signed-off-by: John Mulligan <jmulligan@redhat.com>