cephadm: use get_container_image_stats in cephadm.py
Replace the existing get_container_stats_by_image_name with a version
from container_engines.py that returns the parsed results of the container
status command (as a ContainerInfo, None on error).

Fix up a few tests. Part of the test is somewhat pointless now as the
input and the output are both the same ContainerInfo, but this was never
a great test to test parsing anyway.

Signed-off-by: John Mulligan <jmulligan@redhat.com>