buffer: introduce thin bufferlist_rw to handle iter::copy_in()
The problem the new type tackles is the `copy_in()` on iterator.
This method, per the implicit contract established by its sole
2 users, prohibits invalidation of references and other iter
instances. Therefore we cannot change the structure of underlying
bufferlist to e.g. deep-copy deduplicated zeros; these writes are
truly about touching the content. However, this is just a special
case of far broader problem: crosstalks with regular, ro lists
and pointers.

`bufferlist_rw` deals with that by accepting lists and ptrs only
in the read-write flavours.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>