os: Introduce WITH_KSTORE
Previosuly, we used WITH_SEASTAR to not include KStore with Crimson.
As now Crimson's Alienstore is not defined WITH_SEASTAR We need to be able
to not include KStore to alien crimson target as well.
Instead, let Classic define PRIVATE WITH_KSTORE that will not
be used by the alien targets.

This doesn't change anything of non-crimson targets.

Signed-off-by: Matan Breizman <mbreizma@redhat.com>