mgr/cephadm: fix typo with vrrp_interfaces in keepalive setup
This was intended to be vrrp_interfaces, the variable actually
used later in the code. Instead, due to a typo, it was setting
a variable that is unused other than in the log line right after
it is set. Issue was introduced in
https://github.com/ceph/ceph/commit/58ddc4e20f7cead1f2594241450f4beb5230c746

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit b16fc92155e3f4bbf161610490b579e950584bbf)