cephadm: configure flake8 to ignore E203
The black formatting tool and flake8 linter's default settings disagree
about how to handle spacing around ':' characters in slices.
Use your favorite search engine with the terms "python black flake8
E203" and see various discussions of this topic. Many seem to conclude
that it is E203 that is not pep8 compliant. Go with the flow and just
disable E203 so we can continue to use `format-black` without extra
additional thought :-).

Signed-off-by: John Mulligan <jmulligan@redhat.com>