os/bluestore: Create conf bluestore_onode_segmentation
This debug grade conf selects bluestore_onode_t v2 or v3.
In v2 mode it can read v3 but drops onode_segmentation and pretends
it does not exist. It will act as a version that does not handle v3.
Also it will write onodes in v2 back to DB, clearing segmentation.

Signed-off-by: Adam Kupczyk <akupczyk@ibm.com>