librbd: fix mirror image status summary in a namespace
For the purposes of the summary with image counts, "rbd mirror pool
status" command is supposed to count each image only once.  To this
end, for unidirectional mirroring the status of the receiving site
should be taken while for bidirectional mirroring the statuses should
be combined/reduced.  For example, if mirroring is enabled on a single
image and everything is in order, the summary is expected to be

  image health: OK
  images: 1 total
      1 replaying

on both clusters even though on the primary the local status is
MIRROR_IMAGE_STATUS_STATE_STOPPED and only on the secondary it's
MIRROR_IMAGE_STATUS_STATE_REPLAYING.

Currently this isn't the case for custom namespaces.  In the same
scenario the primary ends up reporting

  image health: OK
  images: 1 total
      1 stopped

based solely on the local status in a namespace.

Fixes: https://tracker.ceph.com/issues/69911
Signed-off-by: Ilya Dryomov <idryomov@gmail.com>
(cherry picked from commit f5eadfff807a84ff6561da3239ee804e735d64b0)

Conflicts:
	qa/workunits/rbd/rbd_mirror_bootstrap.sh [ commits 3fd8a0388735
	  ("qa/workunits/rbd: merge journal and snapshot test scripts")
	  and 3fdbc160bb21 ("rbd-mirror: allow mirroring to a different
	  namespace") not in reef ]