rgw: add dedicated config for max aio during bucket index cleanup
When a bucket with many shards is resharded, especially with large
omaps, the deletion of old index shards can put significant load on
the index pool. The default value of rgw_bucket_index_max_aio (128)
results in deleting 128 shards simultaneously, which can exacerbate
the issue. Introducing a dedicated config to control the maximum
number of deletions allows for better management of this load during
cleanup.

Fixes: https://tracker.ceph.com/issues/69910
Signed-off-by: Seena Fallah <seenafallah@gmail.com>