script/run-make: bump up to clang 19
Supporting new stable compilers releases early on
should make it easier in the future by preventing introduction of
deperacted code.
As this PR already upgrades to 16 with the relevant fixes, we can
use this opportunity to upgrade to the last stable release available.

Note: Most distinct change here is that -Wvla-extension enabled
      by default and will result in lots of warnings.

Signed-off-by: Matan Breizman <mbreizma@redhat.com>