qa/tasks: improve ignorelist for thrashing OSDs
This yaml file is used in rados/thrash-old-clients.
In this commit, I added some pattern matching for
warnings that show up in the cluster log detail that
are related to degraded PGs. In these tests, we are intentionally
marking down or killing OSDs, which leads to these states
showing up in the cluster log. So, the warnings are intended
and can be ignored in the context of OSD thrashing.

This directly aligns the squid ignorelist with main's, rather than
just a direct backport of
https://github.com/ceph/ceph/pull/61723/commits/ec50fc720c4fc0bdd3165d0014f19c36e9819012.

Fixes: https://tracker.ceph.com/issues/69962
Signed-off-by: Laura Flores <lflores@ibm.com>