[ceph-exporter] Fix instanceid extraction logic
Currently the ceph-exporter extracts RGW instance_id from the
RGW daemon socket file name. The socket file is named as:
ceph-client.rgw.<service_name>.<random 6 char string>.<floating_value>.asok
Examples of rgw socket file path:
1. ceph-client.rgw.anmolrgw.ceph-node.bfdfzq.2.187651512351080.asok
2. ceph-client.rgw.realm.zone.ceph01.smfgpn.2.187651512351080.asok
3. ceph-client.rgw.reg.www.aaaaa-ceph-cluster-bootstrap.baawsj.2.187651512351080.asok
Until this PR, the code did the following:
1. Remove the prefixes ceph-client and rgw
2. Extracted string index from the string broken on `.` from string in 1
3. Used string in 2 as instance_id
as can be seen, the above logic results in faulty extraction of instance_id as below:
1. Example 1, bfdfzq
2. Example 2 returns, ceph01
3. Example 3 returns baawsj
The above resulted in an issue where customer had their RGW instance id extracted as ceph01
both pre and post-upgrades although the daemon's 6 char string had changed. This resulted
in multiple combinations of meta-data rows for the same RGW instance and caused many-to-many
joins in the RGW grafana dashboards causing a break in the dashboard.

This PR aims to retain the uniqueness of the daemon instance ids by sanitizing the socket filename
as:
1. Remove standard prefixes ceph-client and rgw
2. Remove the trailing numeric values
3. Use the remainder of the string as instance_id
The new approach is consistent with the way its done in other ceph code blocks.

fixes: https://tracker.ceph.com/issues/69589
Signed-off-by: Anmol Babu <anmolbabu@Anmols-MacBook-Pro.local>