mds/rejoin: Don't fetch the dir is already complete during rejoin
In MDCache::open_undef_inodes_dirfrags, it can so happen that
the same dir is added twice to fetch queue causing it to fetch
twice and hit the assert that dir is not complete for the
second fetch. So don't fetch if the dir is complete.

Fixes: https://tracker.ceph.com/issues/54205
Signed-off-by: Kotresh HR <khiremat@redhat.com>
(cherry picked from commit a069e7a6ac84424a92d059e1b1f6edb4efb36265)