osd: truncate shard object to correct size with ECSwitch.
The implementation of ECSwitch above dropped an override of rollback_append. This was causing rollback to be truncated to the incorrect size.

This fix updates PGBackend to use a replacement API (object_size_to_shard_size) to determine the correct shard size.

This is clearer than the previous override code and is borrowed from the forthcoming EC optimisations.

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>